refactor/be/#158 common에 있던 S3 controller, service를 api로 이동 및 aws패키지로 redis와 결합
