+++ api/build.gradle
@@ -57,6 +57,10 @@ dependencies {
     //redis setting
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 
+    //aws
+    implementation group: 'software.amazon.awssdk', name: 's3', version: '2.31.48'
+    implementation group: 'software.amazon.awssdk', name: 'auth', version: '2.31.48'
+
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
