+++ api/src/main/java/kernel/maidlab/api/aws/service/RedisServiceImpl.java
@@ -1,0 +1,5 @@
+package kernel.maidlab.api.aws.service;
+
+public class RedisServiceImpl {
+
+}
