--- api/src/main/java/kernel/maidlab/api/matching/service/RedisServiceImpl.java
@@ -1,5 +1,0 @@
-package kernel.maidlab.api.matching.service;
-
-public class RedisServiceImpl {
-
-}
