--- common/src/main/java/kernel/maidlab/common/entity/Base/UserBase.java
@@ -1,4 +1,0 @@
-package kernel.maidlab.common.entity.Base;
-
-public interface UserBase {
-}
