+++ common/src/main/java/kernel/maidlab/common/entity/base/UserBase.java
@@ -1,0 +1,4 @@
+package kernel.maidlab.common.entity.base;
+
+public interface UserBase {
+}
