+++ common/src/main/java/kernel/maidlab/common/entity/board/Board.java
@@ -3,10 +3,10 @@ package kernel.maidlab.common.entity.board;
 import jakarta.persistence.*;
 import kernel.maidlab.common.entity.consumer.Consumer;
 import kernel.maidlab.common.entity.manager.Manager;
-import kernel.maidlab.common.entity.Base.UserBase;
+import kernel.maidlab.common.entity.base.UserBase;
 import kernel.maidlab.common.dto.board.request.BoardRequestDto;
 import kernel.maidlab.common.dto.board.request.BoardUpdateRequestDto;
-import kernel.maidlab.common.entity.Base.Base;
+import kernel.maidlab.common.entity.base.Base;
 import kernel.maidlab.common.enums.BoardType;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
