+++ common/src/main/java/kernel/maidlab/common/entity/reservation/Settlement.java
@@ -9,7 +9,7 @@ import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.Table;
-import kernel.maidlab.common.entity.Base.Base;
+import kernel.maidlab.common.entity.base.Base;
 import kernel.maidlab.common.enums.ServiceType;
 import kernel.maidlab.common.enums.Status;
 import lombok.AccessLevel;
