+++ common/src/main/java/kernel/maidlab/common/util/ServletResponseUtil.java
@@ -3,7 +3,7 @@ package kernel.maidlab.common.util;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import jakarta.servlet.http.HttpServletResponse;
-import kernel.maidlab.common.exception.dto.ErrorResponseDto;
+import kernel.maidlab.common.dto.ErrorResponseDto;
 import kernel.maidlab.common.enums.ResponseType;
 import lombok.extern.slf4j.Slf4j;
